Refactor run_check to utils, update setup for CLI, and expand README with nested functionality.
